+++ packages/core-js/stage/3.js
@@ -2,6 +2,7 @@
 var parent = require('./4');
 
 require('../proposals/decorator-metadata-v2');
+require('../proposals/joint-iteration');
 require('../proposals/json-parse-with-source');
 require('../proposals/map-upsert-v4');
 // TODO: Obsolete versions, remove from `core-js@4`